Cleaning up
Apparently the number of transfers increased.
I'll have to look into that, but first let's try the god mode.
